feat(aerosol): wire real MAM4-JAX microphysics core (#490)
Replace the placeholder κ-Köhler core with the real MAM4-JAX modal box
model (reflective-org/MAM4-JAX v0.2.0-beta.1): one step = calcsize →
wateruptake → amicphys (gas–aerosol exchange, rename, binary H2SO4
nucleation, coagulation). Adds Mam4JaxMicrophysics + a full tracer adapter
(m_/mc_/n_/nc_ ↔ MAM4 q/qqcw[pcnst=35] via the data.py index tables) and
fills _jam_state from the core's dgncur_a/dgncur_awet/wetdens. Coherent
with the existing harness: ARG activation, Stokes sedimentation, Slinn dry
dep, wet scavenging and online optics all consume _jam_state unchanged.

The box-model core is vmapped per grid cell. amicphys's implicit diffrax
solver otherwise builds a Jacobian coupled across every cell and the T21
compile exceeds 80 GB; per-cell vmap (each cell integrates the ODE to its
own tolerance — the box model's true semantics) drops T21 to ~4 GB.

mam4-jax is GPL-3.0, so it is an opt-in extra (pip install jcm[mam4])
imported lazily — a plain jcm import stays Apache-2.0. The core only
converges in float64, so the term enables jax_enable_x64; jcm's dtype=float
then runs the whole model in float64 when the core is active. To keep the
model float-structure agnostic, hardcoded jnp.float32 in the Tiedtke-Nordeng
convection scheme (plume + tendency init, downdraft wet-bulb) and a Hines
GWD column buffer now follow the input dtype (no change to float32 runs).

+++ jcm/physics/gravity_waves/hines/hines.py
@@ -477,8 +477,10 @@ def _hines_extro_column(
         2.0 * sqamp_launch / (m_alpha_launch ** 2 - m_min_sq))
     m_alpha_min_running_launch = m_alpha_launch                    # (8,)
 
-    # Initialise full column buffers, indexed top..bottom.
-    m_alpha = jnp.full((nlev, num_azimuths), m_min)
+    # Initialise full column buffers, indexed top..bottom. The fill dtype
+    # follows the computed launch values (not ``m_min``'s) so the buffer is
+    # float-structure agnostic and the scatter below doesn't down/up-cast.
+    m_alpha = jnp.full((nlev, num_azimuths), m_min, dtype=m_alpha_launch.dtype)
     m_alpha = m_alpha.at[launch_idx].set(m_alpha_launch)
     sigma_alpha = jnp.zeros((nlev, num_azimuths))
     sigma_alpha = sigma_alpha.at[launch_idx].set(sigma_a_launch)
